% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_ate_post}
\alias{get_ate_post}
\title{A helper function for getting the posterior of the ATE}
\usage{
get_ate_post(fitted_model, outcome_number, train_or_test)
}
\arguments{
\item{fitted_model}{The fitted mvbcf model to get the ATE posterior from}

\item{outcome_number}{The outcome variable to the ATE posterior for}

\item{train_or_test}{Indicate if the posterior should be for the "train" or "test" data}
}
\value{
The posterior of the ATE for the specified outcome variable
}
\description{
A helper function for getting the posterior of the ATE
}
