% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_mu_preds}
\alias{get_mu_preds}
\title{A helper function for getting mu predictions}
\usage{
get_mu_preds(fitted_model, outcome_number, train_or_test)
}
\arguments{
\item{fitted_model}{The fitted mvbcf model to get predictions from}

\item{outcome_number}{The outcome variable to get predictions for}

\item{train_or_test}{Indicate if predictions should be for the "train" or "test" data}
}
\value{
A vector of individual mu (prognostic effect) estimates for the specified outcome
}
\description{
A helper function for getting mu predictions
}
