% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_sigma_post}
\alias{get_sigma_post}
\title{A helper function for getting the posterior of specified sigma values. If
the first_outcome and second_outcome are the same, then the residual variance
for that outcome is returned. Otherwise the residual covariance for the
first_outcome and second_outcome is returned.}
\usage{
get_sigma_post(fitted_model, first_outcome, second_outcome)
}
\arguments{
\item{fitted_model}{The fitted mvbcf model to get the posterior from}

\item{first_outcome}{The first outcome variable}

\item{second_outcome}{The second outcome variable}
}
\value{
The posterior of the specified residual variance or covariance
}
\description{
A helper function for getting the posterior of specified sigma values. If
the first_outcome and second_outcome are the same, then the residual variance
for that outcome is returned. Otherwise the residual covariance for the
first_outcome and second_outcome is returned.
}
