% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_tau_preds}
\alias{get_tau_preds}
\title{A helper function for getting tau predictions}
\usage{
get_tau_preds(fitted_model, outcome_number, train_or_test)
}
\arguments{
\item{fitted_model}{The fitted mvbcf model to get predictions from}

\item{outcome_number}{The outcome variable to get predictions for}

\item{train_or_test}{Indicate if predictions should be for the "train" or "test" data}
}
\value{
A vector of individual treatment effect estimates for the specified outcome
}
\description{
A helper function for getting tau predictions
}
