% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_y_preds}
\alias{get_y_preds}
\title{A helper function for getting y predictions}
\usage{
get_y_preds(fitted_model, outcome_number, train_or_test, Z)
}
\arguments{
\item{fitted_model}{The fitted mvbcf model to get predictions from}

\item{outcome_number}{The outcome variable to get predictions for}

\item{train_or_test}{Indicate if predictions should be for the "train" or "test" data}

\item{Z}{A binary treatment status indicator: 1 for treatment 0 for control}
}
\value{
A vector of individual y estimates for the specified outcome
}
\description{
A helper function for getting y predictions
}
