% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvbcf_function.R
\name{run_mvbcf}
\alias{run_mvbcf}
\title{The main function for fitting multivariate bcf}
\usage{
run_mvbcf(
  X_con,
  y,
  Z,
  X_mod,
  X_con_test,
  X_mod_test,
  alpha,
  beta,
  alpha_tau,
  beta_tau,
  sigma_mu,
  sigma_tau,
  v_0,
  sigma_0,
  n_iter,
  n_tree,
  n_tree_tau,
  min_nodesize
)
}
\arguments{
\item{X_con}{The control variables used in the mu trees}

\item{y}{The outcome variable/s}

\item{Z}{The treatment indicator}

\item{X_mod}{The effect moderators used in the tau trees}

\item{X_con_test}{The control variables used in the mu trees (test data)}

\item{X_mod_test}{The effect moderators used in the tau trees (test data)}

\item{alpha}{The alpha parameter in the tree prior for mu trees}

\item{beta}{The beta parameter in the tree prior for mu trees}

\item{alpha_tau}{The alpha parameter in the tree prior for tau trees}

\item{beta_tau}{The beta parameter in the tree prior for tau trees}

\item{sigma_mu}{The prior for the terminal node parameters of mu trees}

\item{sigma_tau}{The prior for the terminal node parameter of tau trees}

\item{v_0}{The treatment indicator}

\item{sigma_0}{The effect moderators used in the tau trees}

\item{n_iter}{The number of MCMC iterations}

\item{n_tree}{The number of mu trees}

\item{n_tree_tau}{The number of tau trees}

\item{min_nodesize}{Moves resulting in nodes with fewer observations than this are rejected}
}
\value{
A list of model outputs
}
\description{
The main function for fitting multivariate bcf
}
